+++ addon/components/range-slider.js
@@ -95,23 +95,6 @@ export default Component.extend({
         });
       }
     });
-
-    /** DEPRECATED AND WILL BE REMOVED BEFORE 1.0 **/
-    slider.on('change', () => {
-      run(this, function () {
-          let val = this.get("slider").get();
-          this.sendDeprecatedAction("change", val);
-      });
-    });
-
-    if (!isEmpty(this.get('slide'))) {
-      slider.on('slide', () => {
-        run(this, function () {
-          let val = this.get("slider").get();
-          this.sendDeprecatedAction('slide', val);
-        });
-      });
-    }
   },
 
   didUpdateAttrs() {
@@ -162,16 +145,5 @@ export default Component.extend({
     if (this.get('disabled')) {
       return true;
     }
-  }),
-  /**
-   * Perform a naive check to see if the deprecated action name exists in our
-   * attrs and then log a deprecation warning and trigger the old action.
-   */
-  sendDeprecatedAction(action, value) {
-    let actionName = this.get(`attrs.${action}`);
-    if(!isEmpty(actionName)) {
-      Ember.Logger.warn(`DEPRECATION (ember-cli-nouislider): "${action}" action is deprecated in favor of "on-${action}". Support for "${action}" will be dropped in 1.0`);
-      this.sendAction(action, value);
-    }
-  }
+  })
 });